feat: better AI labels based on custom name, model, and endpoint instead of  `ChatGPT`
